Json->Xml
改换Xml形式完毕!

+++ Assets/Csv/Skill/SpawnObject.cs
@@ -4,16 +4,16 @@ using Csv;
 
 namespace Csv.Skill
 {
-	public class SpawnObject : Csv.Skill.Action
+	public  class SpawnObject : Csv.Skill.Action
 	{
 		/// <summary>
 		/// 子物体ID
 		/// <summary>
-		public readonly int Id;
+		public readonly float Id;
 		/// <summary>
 		/// 子物体类型
 		/// <summary>
-		public readonly int SpawnType;
+		public readonly float SpawnType;
 		/// <summary>
 		/// 子物体存活时间
 		/// <summary>
@@ -21,8 +21,8 @@ namespace Csv.Skill
 
 		public SpawnObject(DataStream data) : base(data)
 		{
-			this.Id = data.GetInt();
-			this.SpawnType = data.GetInt();
+			this.Id = data.GetFloat();
+			this.SpawnType = data.GetFloat();
 			this.Life = data.GetFloat();
 		}
 	}
